+++ CHANGELOG.md
@@ -6,7 +6,7 @@ El formato esta basado[Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 ## [0.0.21] - 2025-10-10
 
 ### Added
-- Se crea  la clase `MoveCalculator` y sus metodos.
+- Se crea  la clase `MoveCalculator`,sus metodos y sus test.
 
 ---